fix(docs): banner
